+++ Sources/Instructions/Core/Public/CoachMarksController.swift
@@ -120,7 +120,7 @@ public extension CoachMarksController {
         controllerWindow = parentViewController.view.window
         coachMarksWindow = coachMarksWindow ?? InstructionsWindow(frame: UIScreen.main.bounds)
 
-        coachMarksViewController.attach(to: coachMarksWindow!)
+        coachMarksViewController.attach(to: coachMarksWindow!, of: parentViewController)
 #endif
         flow.startFlow(withNumberOfCoachMarks: numberOfCoachMarks)
     }
@@ -171,7 +171,9 @@ extension CoachMarksController: Snapshottable {
 
 extension CoachMarksController: OverlayManagerDelegate {
     func didReceivedSingleTap() {
-        flow.showNextCoachMark()
+        if delegate?.shouldHandleOverlayTap(in: self, at: flow.currentIndex) ?? true {
+            flow.showNextCoachMark()
+        }
     }
 }